.dot() and .cross() polished extensively, mag() implemented for vector class.
Lots of tests rewritten.  .__eq__(), .__add__(), .__sub__(), .__neg__(),
implemented for Vector class.  Not sure if the Vector Class plays well with
other sympy expressions, and I need to figure out how to use ._args to make it
work well.

the cross and dot functions are now just wrapper functions to the dot and cross
methods of the UnitVector and Vector classes.